Start reverting to Pablo's suggestions.
